+++ py/uchuu_to_dat.py
@@ -149,8 +149,8 @@ def pre_process_uchuu(in_filepath: str, mode: int, outname_base: str, app_mag_cu
     # the vmax should be calculated from un-k-corrected magnitudes
     V_max = get_max_observable_volume(abs_mag_me, z_eff, app_mag_cut, FOOTPRINT_FRAC)
 
-    colors = np.zeros(count, dtype=np.int8) # TODO compute colors. Use color cut as per Alex's paper.
-    chi = np.zeros(count, dtype=np.int8) # TODO compute chi
+    colors = np.zeros(count, dtype=np.int32) # TODO compute colors. Use color cut as per Alex's paper.
+    chi = np.zeros(count, dtype=np.int32) # TODO compute chi
 
     # Output files
     t1 = time.time()
